Complete drag/drop implementation supporting standard and standalone modes.
