+++ source/client/components/CVMediaManager.ts
@@ -1,6 +1,6 @@
 /**
  * 3D Foundation Project
- * Copyright 2019 Smithsonian Institution
+ * Copyright 2023 Smithsonian Institution
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,6 +20,13 @@ import Notification from "@ff/ui/Notification";
 import CVStandaloneFileManager from "./CVStandaloneFileManager";
 import CVAssetManager from "./CVAssetManager";
 import resolvePathname from "resolve-pathname";
+import ExplorerApplication from "client/applications/ExplorerApplication";
+import MainView from "client/ui/story/MainView";
+import CVDocumentProvider from "./CVDocumentProvider";
+import ImportMenu from "client/ui/story/ImportMenu";
+import CVModel2 from "./CVModel2";
+import { EDerivativeUsage } from "client/schema/model";
+import CSelection from "client/../../libs/ff-graph/source/components/CSelection";
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -80,10 +87,11 @@ export default class CVMediaManager extends CAssetManager
     uploadFile(name: string, blob: Blob, folder: IAssetEntry): Promise<any>
     {
         const filename = decodeURI(name);
-        const url = resolvePathname(folder.info.path + filename, this.rootUrl);
+        const filepath = folder.info.path.length > 1 ? folder.info.path + filename : filename;
+        const url = resolvePathname(filepath, this.rootUrl);
         
         if(this.standaloneFileManager) {
-            this.standaloneFileManager.addFile(CVMediaManager.articleFolder + "/" + filename, [blob]);
+            this.standaloneFileManager.addFile(filepath, [blob]);
             this.refresh();
             return Promise.resolve();
         }
@@ -95,14 +103,82 @@ export default class CVMediaManager extends CAssetManager
 
     ingestFiles(files: Map<string, File>)
     {
-        if(this.standaloneFileManager) {
-            this.standaloneFileManager.onFileDrop(files);
+        // If a scene file has been dropped, push to end
+        const fileArray = Array.from(files);
+        const docIndex = fileArray.findIndex( (element) => { return element[0].toLowerCase().indexOf(".svx.json") > -1 });
+        const documentProvided : boolean = docIndex > -1;
+        if(documentProvided) {
+            fileArray.push(fileArray.splice(docIndex,1)[0]);
+
+            // we have a new scene, so clear out standalone file manager
+            if(this.standaloneFileManager) {
+                this.standaloneFileManager.reload();
+            }
         }
-        else {
-            const fileList = new DataTransfer();
-            files.forEach(file => fileList.items.add(file));
-            this.uploadFiles(fileList.files, this.root);
-        }
+
+        const documentRoot = documentProvided ? fileArray[fileArray.length-1][0].replace(fileArray[fileArray.length-1][1].name, '') : "";
+
+        fileArray.forEach(([path, file]) => {
+            const cleanfileName = decodeURI(file.name);
+            const filenameLower = cleanfileName.toLowerCase();
+            
+            if (filenameLower.match(/\.(gltf|glb|bin|svx.json|html|jpg|png|usdz)$/)) {
+
+                if(!documentProvided && filenameLower.match(/\.(jpg|png)$/) && !fileArray.some(entry => entry[0].endsWith("gltf"))) {
+                    path = CVMediaManager.articleFolder + "/" + cleanfileName;
+                }
+
+                // normalize path relative to document root
+                let normalizedPath = documentProvided ? path.replace(documentRoot, '') : path;
+                normalizedPath = normalizedPath.startsWith("/") ? normalizedPath.substr(1) : normalizedPath;
+
+                if (filenameLower.match(/\.(svx.json)$/)) {
+                    const mainView : MainView = document.getElementsByTagName('voyager-story')[0] as MainView;
+                    const explorer : ExplorerApplication = mainView.app.explorerApp;
+            
+                    this.uploadFile(normalizedPath, file, this.root).then(() => {
+                        explorer.loadDocument(normalizedPath);
+                        this.getMainComponent(CVDocumentProvider).refreshDocument();
+                    }); 
+                }
+                else if (!documentProvided && filenameLower.match(/\.(gltf|glb)$/)) {
+                    this.uploadFile(normalizedPath, file, this.root).then(() => this.handleModelImport(normalizedPath));
+                }
+                else {
+                    this.uploadFile(normalizedPath, file, this.root);
+                }
+            }
+            else {
+                new Notification(`Unhandled file: '${cleanfileName}'`, "warning", 4000);
+            }
+        });
+    }
+
+    protected handleModelImport(filepath: string) {
+        const mainView : MainView = document.getElementsByTagName('voyager-story')[0] as MainView;
+        const activeDoc = this.getMainComponent(CVDocumentProvider).activeComponent;
+        const filename = filepath.substr(filepath.lastIndexOf("/") + 1);
+        const selection = this.getMainComponent(CSelection);
+
+        ImportMenu.show(mainView, activeDoc.setup.language, filename).then(([quality, parentName]) => {
+            const model = this.getSystemComponents(CVModel2).find(element => element.node.name === parentName);
+            if(model === undefined) {
+                // converting path to relative (TODO: check if all browsers will have leading slash here)
+                const newModel = activeDoc.appendModel(filepath, quality);
+                const name = parentName;
+                newModel.node.name = name;
+                newModel.ins.name.setValue(name);
+                newModel.ins.quality.setValue(quality);
+                selection.selectNode(newModel.node);
+            }
+            else {
+                model.derivatives.remove(EDerivativeUsage.Web3D, quality);
+                model.derivatives.createModelAsset(filepath, quality)
+                model.ins.quality.setValue(quality);
+                model.outs.updated.set();
+                selection.selectNode(model.node);
+            }
+        }).catch(e => {});
     }
 
     refresh()
@@ -170,7 +246,7 @@ export default class CVMediaManager extends CAssetManager
     refreshRoot()
     {
         if(this.assetManager.ins.baseUrlValid.value) {
-            super.refresh().then(() => this.rootUrlChanged());
+            this.refresh().then(() => this.rootUrlChanged());
         }
     }
 }
