+++ source/client/components/CVStandaloneFileManager.ts
@@ -1,6 +1,6 @@
 /**
  * 3D Foundation Project
- * Copyright 2019 Smithsonian Institution
+ * Copyright 2023 Smithsonian Institution
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,17 +20,9 @@ import Component, { Node } from "@ff/graph/Component";
 import CVAssetManager from "./CVAssetManager";
 import CVDocumentProvider from "./CVDocumentProvider";
 
-import Notification from "@ff/ui/Notification";
-
-//import { SimpleDropzone } from 'simple-dropzone';
-import ExplorerApplication from "client/applications/ExplorerApplication";
-import MainView from "client/ui/story/MainView";
 import CVMediaManager from "./CVMediaManager";
 import { IFileInfo } from "client/../../libs/ff-scene/source/components/CAssetManager";
 import { Dictionary } from "client/../../libs/ff-core/source/types";
-import ImportMenu from "client/ui/story/ImportMenu";
-import CVModel2 from "./CVModel2";
-import { EDerivativeUsage } from "client/schema/model";
 import CSelection from "client/../../libs/ff-graph/source/components/CSelection";
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -44,11 +36,6 @@ export default class CVStandaloneFileManager extends Component
     private runtimeURLs: Array<string> = [];
     private fileMap: Dictionary<File> = {};
     private rootMap = {};
-    private importQueue: string[] = [];
-
-
-    private isConfigured: boolean = null;
-    private documentLoaded: boolean = false;
 
     protected get documentProvider() {
         return this.getMainComponent(CVDocumentProvider);
@@ -76,6 +63,28 @@ export default class CVStandaloneFileManager extends Component
         this.fileMap["/"] = null;  // needs base path for asset tree to generate correctly
         this.fileMap["articles/"] = null;
         this.mediaManager.refresh();
+
+        // add url modifier callback to return correct url
+        this.assetManager.loadingManager.setURLModifier( ( url ) => {
+            
+            url = decodeURI(url);
+            const index = url.lastIndexOf('/');
+            const filename = url.substr(index + 1).replace(/^(\.?\/)/, '');
+            
+            const baseURL = this.rootMap[filename] + filename;
+            const normalizedURL = baseURL.startsWith("/") ? baseURL.substr(1) : baseURL;  // strip potential leading slash
+            
+            if(this.runtimeURLs[normalizedURL]) {
+                return this.runtimeURLs[normalizedURL];
+            }
+            if(this.fileMap[normalizedURL]) {
+                const bloburl = URL.createObjectURL( this.fileMap[normalizedURL] );
+                this.runtimeURLs[normalizedURL] = bloburl;
+                return bloburl;
+            } 
+            
+            return url;
+        } );
     }
 
     dispose()
@@ -101,16 +110,28 @@ export default class CVStandaloneFileManager extends Component
     
     addFile(uri: string, data: BlobPart[] = [])
     {
+        // add folders to map
+        const parts = uri.split("/").filter(part => !!part);
+        let folderPath = "";
+        for (let i = 0, nj = parts.length-1; i < nj; i++) {
+            folderPath = folderPath.concat(parts[i], "/");
+            this.fileMap[folderPath] = null;
+        }
+
         const idx = uri.lastIndexOf("/");
         const filename = idx > -1 ? uri.substr(idx + 1) : uri;
         const rootPath = uri.replace(filename, '');
+        
+        const normalizedURL = uri.startsWith("/") ? uri.substr(1) : uri;  // strip potential leading slash
 
         this.rootMap[filename] = rootPath;
-        this.fileMap[uri] = new File(data, filename);
+        this.fileMap[normalizedURL] = new File(data, filename);
 
-        if(this.runtimeURLs[uri]) {
-            this.runtimeURLs[uri] = null;
+        if(this.runtimeURLs[normalizedURL]) {
+            this.runtimeURLs[normalizedURL] = null;
         }
+
+        this.mediaManager.refresh();
     }
 
     deleteFile(uri: string)
@@ -138,106 +159,17 @@ export default class CVStandaloneFileManager extends Component
         return Object.keys(urlList).find(key => urlList[key] === bloburl);
     }
 
-    update()
-    {
-        /*if(!this.isConfigured) {
-            const explorerPanel = document.getElementsByClassName('sv-explorer-panel')[0];
-            const input = document.querySelector('#fileInput');
-            const dropZone = new SimpleDropzone(explorerPanel, input);
-            dropZone.on('drop', ({files}: any) => this.onFileDrop(files));
-            this.isConfigured = true;
-        }*/
+    reload() {
+        this.cleanupFiles();
+        this.runtimeURLs.length = 0;
+        this.fileMap = {};
+        this.rootMap = {};
 
-        return false;
+        this.fileMap["/"] = null;  // needs base path for asset tree to generate correctly
+        this.fileMap["articles/"] = null;
+        this.mediaManager.refresh();
     }
     
-    onFileDrop(files: Map<string, File>)
-    {
-        const fileArray = Array.from(files);
-        const hasDoc = this.documentLoaded;
-        
-        const docIndex = fileArray.findIndex( (element) => { return element[0].toLowerCase().indexOf(".svx.json") > -1 });
-        const documentProvided : boolean = docIndex > -1;
-        if(documentProvided) {
-            fileArray.push(fileArray.splice(docIndex,1)[0]);
-
-            if(hasDoc) {
-                this.reload();
-            }
-
-            this.documentLoaded = true;
-        }
-        const documentRoot = documentProvided ? fileArray[fileArray.length-1][0].replace(fileArray[fileArray.length-1][1].name, '') : "";
-
-        fileArray.forEach(([path, file]) => {
-            const cleanfileName = decodeURI(file.name);
-            const filenameLower = cleanfileName.toLowerCase();
-            if (filenameLower.match(/\.(gltf|glb|bin|svx.json|html|jpg|png|usdz)$/)) {
-
-                if(!documentProvided && filenameLower.match(/\.(jpg|png)$/) && !fileArray.some(entry => entry[0].endsWith("gltf"))) {
-                    path = CVMediaManager.articleFolder + "/" + cleanfileName;
-                }
-
-                // add folders to map
-                const parts = path.split("/").filter(part => !!part);
-                let folderPath = "";
-                for (let i = 0, nj = parts.length-1; i < nj; i++) {
-                    folderPath = folderPath.concat(parts[i], "/");
-                    this.fileMap[folderPath] = null;
-                }
-
-                // normalize path relative to document root
-                path = documentProvided ? path.replace(documentRoot, '') : (path.startsWith("/") ? path.substr(1) : path);
-                const rootPath = path.replace(cleanfileName, '');
-                this.rootMap[cleanfileName] = rootPath;
-                
-                this.fileMap[path] = file;
-
-                this.assetManager.loadingManager.setURLModifier( ( url ) => {
-
-                    url = decodeURI(url);
-                    const index = url.lastIndexOf('/');
-                    const filename = url.substr(index + 1).replace(/^(\.?\/)/, '');
-
-                    const baseURL = this.rootMap[filename] + filename;
-                    const normalizedURL = baseURL.startsWith("/") ? baseURL.substr(1) : baseURL;  // strip potential leading slash
-
-                    if(this.runtimeURLs[normalizedURL]) {
-                        return this.runtimeURLs[normalizedURL];
-                    }
-                    if(this.fileMap[normalizedURL]) {
-                        const bloburl = URL.createObjectURL( this.fileMap[normalizedURL] );
-                        this.runtimeURLs[normalizedURL] = bloburl;
-                        return bloburl;
-                    } 
-
-                    return url;
-                } );
-
-                if (!documentProvided && filenameLower.match(/\.(gltf|glb)$/)) {
-                    this.importQueue.push(path);
-                }
-                else if (filenameLower.match(/\.(svx.json)$/)) {
-                    const mainView : MainView = document.getElementsByTagName('voyager-story')[0] as MainView;
-                    const explorer : ExplorerApplication = mainView.app.explorerApp;
-            
-                    explorer.loadDocument(path);
-                    this.documentProvider.refreshDocument();
-                    
-                    //this.assetReader.getJSON(path).then(data => this.documentProvider.createDocument(data, path))
-                }
-            }
-            else {
-                new Notification(`Unhandled file: '${cleanfileName}'`, "warning", 4000);
-            }
-        });
-
-        if(this.importQueue.length > 0) {
-            this.handleModelImport(this.importQueue.pop());
-        }
-        this.mediaManager.refresh(); 
-    }
-
     protected cleanupFiles() 
     {
         this.runtimeURLs.forEach(( url ) => URL.revokeObjectURL( url ));
@@ -273,47 +205,5 @@ export default class CVStandaloneFileManager extends Component
         info.path = path;
 
         return info as IFileInfo;
-    }
-
-    protected handleModelImport(filepath: string) {
-        const mainView : MainView = document.getElementsByTagName('voyager-story')[0] as MainView;
-        const activeDoc = this.documentProvider.activeComponent;
-        const importQueue = this.importQueue;
-        const filename = this.fileMap[filepath].name;
-
-        ImportMenu.show(mainView, activeDoc.setup.language, filename).then(([quality, parentName]) => {
-            const model = this.getSystemComponents(CVModel2).find(element => element.node.name === parentName);
-            if(model === undefined) {
-                // converting path to relative (TODO: check if all browsers will have leading slash here)
-                const newModel = activeDoc.appendModel(filepath, quality);
-                const name = parentName;
-                newModel.node.name = name;
-                newModel.ins.name.setValue(name);
-                newModel.ins.quality.setValue(quality);
-                this.selection.selectNode(newModel.node);
-            }
-            else {
-                model.derivatives.remove(EDerivativeUsage.Web3D, quality);
-                model.derivatives.createModelAsset(filepath, quality)
-                model.ins.quality.setValue(quality);
-                model.outs.updated.set();
-                this.selection.selectNode(model.node);
-            }
-
-            if(importQueue.length > 0) {
-                this.handleModelImport(importQueue.pop());
-            }
-        }).catch(e => {});
-    }
-
-    protected reload() {
-        this.cleanupFiles();
-        this.runtimeURLs.length = 0;
-        this.fileMap = {};
-        this.rootMap = {};
-
-        this.fileMap["/"] = null;  // needs base path for asset tree to generate correctly
-        this.fileMap["articles/"] = null;
-        this.mediaManager.refresh();
     }
 }
